requirements.txt
torch
torchvision
torchaudio
numpy
pillow
scikit-learn
matplotlib
opencv-python
streamlit
